improve configuration structure
