+++ components/cqrs/event_bus.go
@@ -2,15 +2,72 @@ package cqrs
 
 import (
 	"context"
+	stdErrors "errors"
 
+	"github.com/ThreeDotsLabs/watermill"
 	"github.com/ThreeDotsLabs/watermill/message"
 	"github.com/pkg/errors"
 )
 
+type EventBusConfig struct {
+	GeneratePublishTopic GenerateEventPublishTopicFn
+
+	// OnPublish is called before sending the event.
+	// The *message.Message can be modified.
+	//
+	// This option is not required.
+	OnPublish OnEventSendFn
+
+	// Marshaler is used to marshal and unmarshal events.
+	// It is required.
+	Marshaler CommandEventMarshaler
+
+	// Logger instance used to log.
+	// If not provided, watermill.NopLogger is used.
+	Logger watermill.LoggerAdapter
+}
+
+func (c *EventBusConfig) setDefaults() {
+	if c.Logger == nil {
+		c.Logger = watermill.NopLogger{}
+	}
+}
+
+func (c EventBusConfig) Validate() error {
+	var err error
+
+	if c.Marshaler == nil {
+		err = stdErrors.Join(err, errors.New("missing Marshaler"))
+	}
+
+	if c.GeneratePublishTopic == nil {
+		err = stdErrors.Join(err, errors.New("missing GenerateHandlerTopic"))
+	}
+
+	return err
+}
+
+type GenerateEventPublishTopicFn func(GenerateEventPublishTopicParams) (string, error)
+
+type GenerateEventPublishTopicParams struct {
+	EventName string
+	Event     any
+}
+
+type OnEventSendFn func(params OnEventSendParams) error
+
+type OnEventSendParams struct {
+	EventName string
+	Event     any
+
+	// Message is never nil and can be modified.
+	Message *message.Message
+}
+
 // EventBus transports events to event handlers.
 type EventBus struct {
 	publisher message.Publisher
-	config    EventConfig
+	config    EventBusConfig
 }
 
 // NewEventBus creates a new CommandBus.
@@ -32,7 +89,7 @@ func NewEventBus(
 
 	return &EventBus{
 		publisher: publisher,
-		config: EventConfig{
+		config: EventBusConfig{
 			GeneratePublishTopic: func(params GenerateEventPublishTopicParams) (string, error) {
 				return generateTopic(params.EventName), nil
 			},
@@ -42,13 +99,13 @@ func NewEventBus(
 }
 
 // NewEventBusWithConfig creates a new EventBus.
-func NewEventBusWithConfig(publisher message.Publisher, config EventConfig) (*EventBus, error) {
+func NewEventBusWithConfig(publisher message.Publisher, config EventBusConfig) (*EventBus, error) {
 	if publisher == nil {
 		return nil, errors.New("missing publisher")
 	}
 
 	config.setDefaults()
-	if err := config.ValidateForBus(); err != nil {
+	if err := config.Validate(); err != nil {
 		return nil, errors.Wrap(err, "invalid config")
 	}
